Adjusting code to header moved to include/cpp_common
