+++ include/c_common/vroom/matrix_input.h
@@ -33,8 +33,8 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include <stddef.h>
 
 #include "c_common/get_check_data.h"
-#include "c_types/column_info_t.h"
-#include "c_types/vroom/vroom_matrix_t.h"
+#include "cpp_common/info.hpp"
+#include "cpp_common/vroom_matrix_t.hpp"
 
 #ifdef PROFILE
 #include "c_common/debug_macro.h"
