+++ include/c_common/vroom/time_windows_input.h
@@ -34,8 +34,8 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include <stddef.h>
 
 #include "c_common/get_check_data.h"
-#include "c_types/column_info_t.h"
-#include "c_types/vroom/vroom_time_window_t.h"
+#include "cpp_common/info.hpp"
+#include "cpp_common/vroom_time_window_t.hpp"
 
 #ifdef PROFILE
 #include "c_common/debug_macro.h"
