+++ include/drivers/pgr_pickDeliverEuclidean_driver.h
@@ -39,9 +39,9 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #   include <stddef.h>
 #endif
 
-#include "c_types/vehicle_t.h"
 typedef struct PickDeliveryOrders_t PickDeliveryOrders_t;
 typedef struct Solution_rt Solution_rt;
+typedef struct Vehicle_t Vehicle_t;
 
 #ifdef __cplusplus
 extern "C" {
