+++ include/problem/fleet.hpp
@@ -34,7 +34,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include <numeric>
 
 #include "c_types/typedefs.h"
-#include "c_types/short_vehicle.h"
+#include "cpp_common/short_vehicle.hpp"
 #include "problem/vehicle_pickDeliver.hpp"
 
 typedef struct Vehicle_t Vehicle_t;
