+++ src/common/vehicles_input.c
@@ -31,8 +31,8 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include <float.h>
 
 
-#include "c_types/column_info_t.h"
-#include "c_types/vehicle_t.h"
+#include "cpp_common/info.hpp"
+#include "cpp_common/vehicle_t.hpp"
 
 #include "c_common/get_check_data.h"
