+++ src/compatibleVehicles/compatibleVehicles_driver.cpp
@@ -36,9 +36,9 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 #include "problem/pickDeliver.hpp"
 #include "problem/matrix.hpp"
-#include "c_types/pickDeliveryOrders_t.h"
+#include "cpp_common/orders_t.hpp"
 #include "c_types/compatibleVehicles_rt.h"
-#include "c_types/vehicle_t.h"
+#include "cpp_common/vehicle_t.hpp"
 
 #include "cpp_common/assert.hpp"
 #include "c_common/pgr_alloc.hpp"
