+++ src/cpp_common/base_matrix.cpp
@@ -39,8 +39,8 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 #include "cpp_common/identifiers.hpp"
 #include "cpp_common/assert.hpp"
-#include "c_types/matrix_cell_t.h"
-#include "c_types/vroom/vroom_matrix_t.h"
+#include "cpp_common/matrix_cell_t.hpp"
+#include "cpp_common/vroom_matrix_t.hpp"
 
 
 namespace vrprouting {
