+++ src/optimize/optimize.c
@@ -36,7 +36,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include "c_common/matrixRows_input.h"
 #include "c_common/time_multipliers_input.h"
 #include "c_types/short_vehicle_rt.h"
-#include "c_types/pickDeliveryOrders_t.h"
+#include "cpp_common/orders_t.hpp"
 #include "drivers/optimize_driver.h"
 
 PGDLLEXPORT Datum
