+++ src/pgr_pickDeliver/pickDeliverEuclidean.c
@@ -33,7 +33,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include "c_common/time_msg.h"
 #include "c_common/orders_input.h"
 #include "c_common/vehicles_input.h"
-#include "c_types/pickDeliveryOrders_t.h"
+#include "cpp_common/orders_t.hpp"
 #include "c_types/solution_rt.h"
 
 #include "drivers/pgr_pickDeliverEuclidean_driver.h"
