+++ src/pgr_pickDeliver/pickDeliverEuclidean_driver.cpp
@@ -35,7 +35,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include <vector>
 #include <map>
 
-#include "c_types/pickDeliveryOrders_t.h"
+#include "cpp_common/orders_t.hpp"
 #include "c_types/solution_rt.h"
 #include "c_common/pgr_alloc.hpp"
 #include "cpp_common/assert.hpp"
