+++ src/pickDeliver/pickDeliver_driver.cpp
@@ -36,9 +36,9 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 #include <deque>
 
 #include "problem/pickDeliver.hpp"
-#include "c_types/pickDeliveryOrders_t.h"
+#include "cpp_common/orders_t.hpp"
 #include "c_types/solution_rt.h"
-#include "c_types/vehicle_t.h"
+#include "cpp_common/vehicle_t.hpp"
 #include "problem/matrix.hpp"
 
 #include "cpp_common/assert.hpp"
