+++ src/problem/matrix.cpp
@@ -38,8 +38,8 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 
 #include "cpp_common/assert.hpp"
-#include "c_types/matrix_cell_t.h"
-#include "c_types/time_multipliers_t.h"
+#include "cpp_common/matrix_cell_t.hpp"
+#include "cpp_common/time_multipliers_t.hpp"
 
 
 namespace vrprouting {
