+++ src/problem/tw_node.cpp
@@ -30,8 +30,8 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 #include "cpp_common/assert.hpp"
 #include "problem/matrix.hpp"
-#include "c_types/pickDeliveryOrders_t.h"
-#include "c_types/vehicle_t.h"
+#include "cpp_common/orders_t.hpp"
+#include "cpp_common/vehicle_t.hpp"
 
 
 namespace vrprouting {
