+++ src/vroom/vroom_driver.cpp
@@ -25,7 +25,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
  ********************************************************************PGR-GNU*/
 
-#include "drivers//vroom_driver.h"
+#include "drivers/vroom_driver.h"
 
 #include <sstream>
 #include <vector>
